move debug to retained
