+++ src/retained.rs
@@ -48,7 +48,7 @@ impl<T: TypePath + Send + Sync> Plugin for RetainedRaycastingPlugin<T> {
         #[cfg(feature = "debug")]
         app.add_systems(
             First,
-            crate::debug::update_debug_cursor::<T>
+            debug::update_debug_cursor::<T>
                 .in_set(RaycastSystem::UpdateDebugCursor::<T>)
                 .run_if(|state: Res<RaycastPluginState<T>>| state.update_debug_cursor)
                 .after(RaycastSystem::UpdateIntersections::<T>),
@@ -461,6 +461,50 @@ pub fn update_target_intersections<T: TypePath + Send + Sync>(
                     .push((source_entity, intersection.to_owned()));
                 previously_updated_raycast_meshes.push(*mesh_entity);
             }
+        }
+    }
+}
+
+#[cfg(feature = "debug")]
+pub mod debug {
+    #![allow(unused)]
+
+    use bevy::{prelude::*, reflect::TypePath};
+    use std::marker::PhantomData;
+
+    use crate::prelude::*;
+
+    /// Updates the 3d cursor to be in the pointed world coordinates
+    #[allow(clippy::too_many_arguments)]
+    pub fn update_debug_cursor<T: TypePath + Send + Sync>(
+        mut commands: Commands,
+        mut meshes: Query<&RaycastSource<T>>,
+        mut gizmos: Gizmos,
+    ) {
+        for (is_first, intersection) in meshes.iter().flat_map(|m| {
+            m.intersections()
+                .iter()
+                .map(|i| i.1.clone())
+                .enumerate()
+                .map(|(i, hit)| (i == 0, hit))
+        }) {
+            let color = match is_first {
+                true => Color::GREEN,
+                false => Color::PINK,
+            };
+            gizmos.ray(intersection.position(), intersection.normal(), color);
+            gizmos.circle(intersection.position(), intersection.normal(), 0.1, color);
+        }
+    }
+
+    /// Used to debug [`RaycastMesh`] intersections.
+    pub fn print_intersections<T: TypePath + Send + Sync>(query: Query<&RaycastMesh<T>>) {
+        for (_, intersection) in query.iter().flat_map(|mesh| mesh.intersections.iter()) {
+            info!(
+                "Distance {:?}, Position {:?}",
+                intersection.distance(),
+                intersection.position()
+            );
         }
     }
 }
